+++ Interface/AddOns/ThreatClassic2/locale/enUS.lua
@@ -101,6 +101,7 @@ L.warnings_flash		= "Enable screen flashing."
 L.warnings_soundHeader	= "Sound warning settings."
 L.warnings_sound		= "Enable sounds."
 L.warnings_threshold	= "Warning threshold (100% = aggro)."
+L.warnings_minThreatAmount = "Minimum threat to trigger warnings."
 
 L.warnings_soundFile	= "Sound file"
 L.warnings_soundChannel	= "Sound channel"