+++ Interface/AddOns/ThreatClassic2/locale/zhCN.lua
@@ -29,7 +29,8 @@ L.version_incompatible	= "不兼容！"
 
 L.general				= "一般"
 L.general_welcome		= "显示欢迎信息当你登陆时."
-L.general_update		= "更新间隔时间 (秒)"
+L.general_rawPercent    = "显示原始威胁百分比（近战110％/ 130％范围内的总攻击力）"   ---
+L.general_updateFreq    = "更新间隔的时间（以秒为单位）."   --
 L.general_test			= "测试模式."
 L.general_minimap		= "小地图按钮."
 L.general_ignorePets	= "忽略玩家宠物的仇恨."
@@ -38,11 +39,15 @@ L.visibility			= "可见性"
 L.visibility_hideOOC	= "当脱战时隐藏框架"
 L.visibility_hideSolo	= "当不在队伍中隐藏框架"
 L.visibility_hideInPvP	= "在战场中隐藏框架"
+L.visibility_hideOpenWorld	= "不在团队中隐藏框架 (也隐藏世界BOSS!)."
+L.visibility_hideAlways	= "始终隐藏框架。 /tc2 切换"
 
 L.nameplates			= "姓名版"
 L.nameplates_enable		= "启用仇恨着色 (仅暴雪默认姓名版)."
 L.nameplates_invert		= "颠倒仇恨着色 (便于坦克)."
 L.nameplates_colors		= "仇恨颜色"
+
+L.profiles              = "轮廓"
 
 L.color					= "颜色"
 L.color_good			= "低"
@@ -55,12 +60,16 @@ L.frame					= "框架"
 L.frame_header			= "标题"
 L.frame_bg				= "背景"
 L.frame_test			= "测试模式"
-L.frame_strata			= "Strata"
+L.frame_strata			= "层面"
 L.frame_scale			= "大小"
 L.frame_lock			= "锁定"
 L.frame_headerShow		= "显示标题"
 L.frame_headerColor		= "标题颜色"
-L.frame_width			= "宽度"
+L.frame_position		= "位置"--
+L.frame_width			= "宽"
+L.frame_height			= "高"
+L.frame_xOffset         = "X 偏移"
+L.frame_yOffset         = "Y 偏移"---
 
 L.bar					= "仇恨条"
 L.bar_count				= "最大数量"
@@ -74,15 +83,30 @@ L.bar_defaultColor		= "自定义颜色"
 L.bar_alpha				= "条透明度 (default colors)"
 L.bar_colorMod			= "颜色调节"
 
+L.customBarColors 			        = "自定义颜色"
+L.customBarColorsPlayer_enabled	    = "启用自定义玩家颜色"
+L.customBarColorsActiveTank_enabled	= "启用自定义坦克颜色"
+L.customBarColorsOtherUnit_enabled 	= "启用自定义其他玩家颜色"
+L.customBarColorsPlayer_color 	    = "玩家颜色"
+L.customBarColorsActiveTank_color 	= "主坦克颜色"
+L.customBarColorsOtherUnit_color 	= "其他玩家颜色"
+
 L.font					= "字体"
-L.font_name			= "名字"
+L.font_name			    = "字体"
 L.font_size				= "大小"
 L.font_style			= "风格"
 L.font_shadow			= "字体阴影"
 
 L.warnings				= "警报"
+L.warnings_flash		= "启用屏幕闪烁."   ---
 L.warnings_sound		= "启用声音."
 L.warnings_threshold	= "警报仇恨阈值 (100% = aggro)."
 
 L.warnings_soundFile		= "警报声音文件."
-L.sound_pulledFile		= "获得仇恨警报的声音文件."
+L.warnings_soundChannel	= "声音通道"
+
+L.soundChannel_master = "主"
+L.soundChannel_sfx = "特效"
+L.soundChannel_ambience = "环境"
+L.soundChannel_music = "音乐"
+